Migrate parsesSearchQuery to query state store (take 2) (#29631)
This PR restores #29449 by reverting #29565 . See the original PR for information about the actual changes.

Additionally this PR includes a fix for what the original PR got wrong. This is all in the second commit so it should suffice to review that:

The issue with the previous implementation was that the route render function wasn't run again after the search was submitted and thus the search results component was never rendered.

This commit fixes this issue by wrapping the conditional logic in an actual component which can properly subscribe to the query state store and gets re-rendered when a search query is present in the query (or not). After chatting with other frontend devs, this seems to be the most sensible solution at this time.

This wasn't a problem in the past because all data was passed down via props, guaranteeing that the parent component would re-render if anything changes. But this is different now that we move things to stores and directly reading data from the stores is prone to such errors.