Clean up CommitMatch string types (#29822)
This pushes the MatchedString type up the stack so that CommitMatch uses MatchedString instead of HighlightedString. This is useful because MatchedString retains multiline information. Additionally, this moves the conversion code from MatchedString to HighlightedString into the internal/search/result package, which makes sense because both of those types are defined there.