search: remove duplicate event SearchResultClicked (#29910)
For each click on a file match in the list of search results, we currently send out 2 event logs
of type "SearchResultClicked" with the only difference being the url (search page vs. blob page).

This removes the call to `onSelect` from `FileMatchChildren` and keeps it for the `ResultContainer`.
This way we can treat the other match types, IE those without children, the same.