git-combine: limit flag for a target commit history size (#30333)
For the gigarepo we want to ensure we start off with a size of around
1.1million commits. We will use this flag to target that size.

Note we already had a limit flag, but that was used to prevent storing too
much information in memory from each remote. This flag still exists and is
useful, but has been renamed to limit-remote.