Search backend: pass stream through explicitly (#30459)
This modifies our search pipeline to pass the stream through each and every step that needs it. The goal here is to remove references to r.stream, especially mutations of r.stream so that we can be sure that either 1) we are running in streaming mode and stream is non-nil, or 2) we are running in batch mode and stream is nil.