ssbc: Don't persist steps and skipped steps on workspaces (#30291)
After recent refactors, we don't modify the steps slice anymore, so the value is just duplicated from what's stored in the associated batch spec. To have a single source of truth, we remove that here.
Also, we can compute the skipped steps for cheap in the application layer, no need to persist this small computation in the database and risk it getting out of date or needing to migrate data later on.
We will make use of this when we carry over workspaces from one batch spec to another when there was no change in the on and workspaces sections.