gitserver: Only allow acceptable commands and POST method in /exec (#30833)
Patch for https://github.com/sourcegraph/security-issues/issues/213.

This feature is guarded behind a feature flag: 
experimentalFeatures.enableGitServerCommandExecFilter

The feature flag is turned off by default and allows us to flip it
when we are confident with the changes in production after observing
the logs. Also, it lets us flip it back to the default in case there
are any regressions.

This commit also adds a test to validate existing behaviour with and
without the feature flag enabled.

Finaly, this commit adds a metric to count the number of blocked
command executions.